+++ tests/test_21_end2end/sbpl-pkg.sh
@@ -2,5 +2,5 @@
 set -eu
 
 if ! command -v 'foo'; then
-    sbpl_get 'archive' 'foo' '1.0.0' 'https://github.com/octocraft/${name}/raw/v${version}/dist/${name}.zip' './bin/${OS}/${ARCH}'
+    sbpl_get 'archive' 'foo' '1.0.0' 'https://github.com/octocraft/${name}/raw/v${version}/dist/${name}.zip' './bin/${sbpl_os}/${sbpl_arch}'
 fi